+++ content/blog/accessibility-best-practices.mdx
@@ -1,14 +1,12 @@
 ---
 title: "Accessibility Best Practices: Building Inclusive Web Experiences"
-excerpt: "A comprehensive guide to creating accessible web applications that work for everyone"
-date: "2024-01-05"
+excerpt: ""
+date: "2022-04-05"
 tags: ["Accessibility", "A11y", "WCAG", "Inclusive Design", "React"]
 author: "Liceth Ovalles"
 featured: true
 readingTime: "10 min read"
 ---
-
-# Accessibility Best Practices: Building Inclusive Web Experiences
 
 Accessibility isn't just about compliance—it's about creating web experiences that work for everyone. In this guide, I'll share practical techniques for building accessible React applications that follow WCAG 2.1 guidelines.
 
@@ -445,8 +443,6 @@ Accessibility is not a feature—it's a fundamental requirement for creating inc
 Remember: **Accessible design is good design.** When you build for accessibility, you're building for all users.
 
 ---
-
-*What accessibility challenges have you faced in your projects? Share your experiences and solutions in the comments!* ♿️
 
 ## Additional Resources
