+++ content/blog/building-kawaii-portfolio.mdx
@@ -1,14 +1,12 @@
 ---
 title: "Building a Kawaii Portfolio: Where Design Meets Performance"
-excerpt: "How I created a delightful, performant portfolio that showcases both technical skills and personality"
-date: "2024-01-15"
+excerpt: ""
+date: "2024-08-17"
 tags: ["React", "Next.js", "Design", "Performance", "Animation"]
 author: "Liceth Ovalles"
 featured: true
 readingTime: "8 min read"
 ---
-
-# Building a Kawaii Portfolio: Where Design Meets Performance
 
 Creating a portfolio that stands out in today's competitive frontend market requires more than just technical skills—it needs personality, performance, and that special touch that makes visitors remember you.
 
@@ -160,4 +158,3 @@ The key is balancing innovation with reliability, personality with professionali
 
 *Building a portfolio is an ongoing journey. The best portfolios evolve with you, showcasing not just what you can do, but who you are as a developer.*
 
-**What makes your portfolio unique? Share your thoughts in the comments below!** 💖
