+++ content/blog/react-performance-tips.mdx
@@ -1,14 +1,12 @@
 ---
 title: "React Performance Tips: From Good to Great"
-excerpt: "Practical techniques to optimize React applications for better user experience and Core Web Vitals"
-date: "2024-01-10"
+excerpt: ""
+date: "2024-09-28"
 tags: ["React", "Performance", "Optimization", "Web Vitals"]
 author: "Liceth Ovalles"
 featured: false
 readingTime: "6 min read"
 ---
-
-# React Performance Tips: From Good to Great
 
 Performance optimization in React isn't just about making your app faster—it's about creating a better user experience that keeps visitors engaged and improves your Core Web Vitals scores.
 
@@ -338,5 +336,3 @@ Performance optimization is an ongoing process, not a one-time task. By implemen
 Remember: **Fast is a feature.** Users expect applications to be responsive and smooth. By prioritizing performance, you're prioritizing user experience.
 
 ---
-
-*What performance techniques have made the biggest impact on your React applications? Share your experiences in the comments!* 🚀
