+++ src/app/components/work/work.tsx
@@ -4,7 +4,6 @@ import { motion } from "framer-motion";
 import { staggerContainer, fadeInUp } from "../../utils/animations";
 import AnimatedText from "../AnimatedText";
 import KawaiiCard from "../KawaiiCard";
-import { Blogs } from "./blogs";
 
 export function Work() {
   const experience = [
@@ -87,9 +86,6 @@ export function Work() {
   ];
   return (
     <section id="work" className="py-5 px-6 max-w-7xl mx-auto">
-      {/* Blog Section */}
-      <Blogs />
-     
       {/* Work Experience */}
       <AnimatedText 
         type="fade"
