+++ src/app/utils/blogUtils.ts
@@ -128,26 +128,11 @@ export function getRelatedPosts(currentSlug: string, limit: number = 3): BlogPos
     .sort((a, b) => b.commonTagsCount - a.commonTagsCount)
     .slice(0, limit);
 
-  return relatedPosts.map(({ commonTagsCount, ...post }) => post);
+  return relatedPosts.map(({  ...post }) => post);
 }
 
-// Format date for display
-export function formatDate(dateString: string): string {
-  const date = new Date(dateString);
-  return date.toLocaleDateString('en-US', {
-    year: 'numeric',
-    month: 'long',
-    day: 'numeric',
-  });
-}
-
-// Generate reading time estimate
-export function estimateReadingTime(content: string): string {
-  const wordsPerMinute = 200;
-  const words = content.split(/\s+/).length;
-  const minutes = Math.ceil(words / wordsPerMinute);
-  return `${minutes} min read`;
-}
+// Re-export utility functions from dateUtils
+export { formatDate, estimateReadingTime } from './dateUtils';
 
 // Search posts by title and content
 export function searchPosts(query: string): BlogPostMeta[] {
